+++ examples/lump-mpi.py
@@ -24,8 +24,6 @@ THE SOFTWARE.
 import logging
 import numpy as np
 import pyopencl as cl
-import numpy.linalg as la  # noqa
-import pyopencl.array as cla  # noqa
 import pyopencl.tools as cl_tools
 from functools import partial
 from mpi4py import MPI
