+++ test/test_euler.py
@@ -43,7 +43,6 @@ from grudge.eager import interior_trace_pair
 from grudge.symbolic.primitives import TracePair
 from mirgecom.euler import inviscid_operator
 from mirgecom.euler import split_conserved
-from mirgecom.steppers import euler_flow_stepper
 from mirgecom.initializers import Vortex2D
 from mirgecom.initializers import Lump
 from mirgecom.boundary import PrescribedBoundary
@@ -53,6 +52,11 @@ from grudge.eager import EagerDGDiscretization
 from pyopencl.tools import (  # noqa
     pytest_generate_tests_for_pyopencl as pytest_generate_tests,
 )
+
+from grudge.shortcuts import make_visualizer
+from mirgecom.euler import get_inviscid_timestep
+from mirgecom.euler import split_fields
+from mirgecom.integrators import rk4_step
 
 
 @pytest.mark.parametrize("dim", [1, 2, 3])
@@ -712,6 +716,115 @@ def test_lump_rhs(ctx_factory, dim, order):
     )
 
 
+def _euler_flow_stepper(actx, parameters):
+    """
+    Implements a generic time stepping loop for testing an inviscid flow.
+    """
+    logging.basicConfig(format="%(message)s", level=logging.INFO)
+    logger = logging.getLogger(__name__)
+
+    mesh = parameters['mesh']
+    t = parameters['time']
+    order = parameters['order']
+    t_final = parameters['tfinal']
+    initializer = parameters['initializer']
+    exittol = parameters['exittol']
+    casename = parameters['casename']
+    boundaries = parameters['boundaries']
+    eos = parameters['eos']
+    cfl = parameters['cfl']
+    dt = parameters['dt']
+    constantcfl = parameters['constantcfl']
+    nstepstatus = parameters['nstatus']
+
+    if t_final <= t:
+        return(0.0)
+
+    rank = 0
+    dim = mesh.dim
+    istep = 0
+
+    discr = EagerDGDiscretization(actx, mesh, order=order)
+    nodes = thaw(actx, discr.nodes())
+    fields = initializer(0, nodes)
+    sdt = get_inviscid_timestep(discr, fields, cfl=cfl, eos=eos)
+
+    initname = initializer.__class__.__name__
+    eosname = eos.__class__.__name__
+    message = (
+        f"Num {dim}d order-{order} elements: {mesh.nelements}\n"
+        f"Timestep:        {dt}\n"
+        f"Final time:      {t_final}\n"
+        f"Status freq:     {nstepstatus}\n"
+        f"Initialization:  {initname}\n"
+        f"EOS:             {eosname}"
+    )
+    logger.info(message)
+
+    vis = make_visualizer(discr, discr.order + 3 if dim == 2 else discr.order)
+
+    def write_soln(write_status=True):
+        dv = eos(fields)
+        expected_result = initializer(t, nodes)
+        result_resid = fields - expected_result
+        maxerr = [np.max(np.abs(result_resid[i].get())) for i in range(dim + 2)]
+        mindv = [np.min(dvfld.get()) for dvfld in dv]
+        maxdv = [np.max(dvfld.get()) for dvfld in dv]
+
+        if write_status is True:
+            statusmsg = (
+                f"Status: Step({istep}) Time({t})\n"
+                f"------   P({mindv[0]},{maxdv[0]})\n"
+                f"------   T({mindv[1]},{maxdv[1]})\n"
+                f"------   dt,cfl = ({dt},{cfl})\n"
+                f"------   Err({maxerr})"
+            )
+            logger.info(statusmsg)
+
+        io_fields = split_fields(dim, fields)
+        io_fields += eos.split_fields(dim, dv)
+        io_fields.append(("exact_soln", expected_result))
+        io_fields.append(("residual", result_resid))
+        nameform = casename + "-{iorank:04d}-{iostep:06d}.vtu"
+        visfilename = nameform.format(iorank=rank, iostep=istep)
+        vis.write_vtk_file(visfilename, io_fields)
+
+        return maxerr
+
+    def rhs(t, q):
+        return inviscid_operator(discr, q=q, t=t, boundaries=boundaries, eos=eos)
+
+    while t < t_final:
+
+        if constantcfl is True:
+            dt = sdt
+        else:
+            cfl = dt / sdt
+
+        if nstepstatus > 0:
+            if istep % nstepstatus == 0:
+                write_soln()
+
+        fields = rk4_step(fields, t, dt, rhs)
+        t += dt
+        istep += 1
+
+        sdt = get_inviscid_timestep(discr, fields, cfl=cfl, eos=eos)
+
+    if nstepstatus > 0:
+        logger.info("Writing final dump.")
+        maxerr = max(write_soln(False))
+    else:
+        expected_result = initializer(t, nodes)
+        maxerr = discr.norm(fields - expected_result, np.inf)
+
+    logger.info(f"Max Error: {maxerr}")
+    if maxerr > exittol:
+        raise ValueError("Solution failed to follow expected result.")
+
+    return(maxerr)
+
+
 @pytest.mark.parametrize("order", [1, 2, 3])
 def test_isentropic_vortex(ctx_factory, order):
     """Advance the 2D isentropic vortex case in
@@ -760,7 +873,7 @@ def test_isentropic_vortex(ctx_factory, order):
                       'eos': eos, 'casename': casename, 'mesh': mesh,
                       'tfinal': t_final, 'exittol': exittol, 'cfl': cfl,
                       'constantcfl': False, 'nstatus': 0}
-        maxerr = euler_flow_stepper(actx, flowparams)
+        maxerr = _euler_flow_stepper(actx, flowparams)
         eoc_rec.add_data_point(1.0 / nel_1d, maxerr)
 
     message = (
